feat: 퀴즈 삭제 기능 리팩토링
